extracted shared logic into helpers and utils
