+++ src/calls/embedTemplateEditorPage.ts
@@ -1,5 +1,10 @@
-import * as z from "@zod/mini";
 import type {BadgeSdk} from "../sdk.ts";
+import type {AppearanceConfig} from "../helpers/appearance.ts";
+import type {FontSource} from "../helpers/appearance.ts";
+import {createEmbedIframe} from "../utils/iframe.ts";
+import {validatePermissions} from "../helpers/permissions.ts";
+import {parseTokenPayload} from "../helpers/token.ts";
+import {type SdkPermission} from "../helpers/permissions.ts";
 
 export interface EmbedTemplateEditorPageOptions {
   /**
@@ -17,149 +22,34 @@ export interface EmbedTemplateEditorPageOptions {
 export function embedTemplateEditorPage(
   sdk: BadgeSdk,
   element: HTMLElement,
-  options: EmbedTemplateEditorPageOptions,
+  {templateId, features, fonts, appearance}: EmbedTemplateEditorPageOptions,
 ): void {
-  const {templateId, features, fonts, appearance} = options;
   const {workspaceHandle, permissions} = parseTokenPayload(sdk.token);
 
-  validatePermissions(permissions, features ?? {});
+  validatePermissions(permissions, REQUIRED_PERMISSIONS);
+  validateFeatures(features ?? {});
 
   const iframe = createEmbedIframe({
     token: sdk.token,
     path: sdk.path,
-    workspaceHandle,
-    templateId,
-    config: {
-      features,
-      fonts,
-      appearance,
-    },
+    pathHash: `/${workspaceHandle}/templates/${templateId}/edit`,
+    config: {features, fonts, appearance},
   });
 
   element.replaceChildren(iframe);
 }
 
-interface CreateIframeOptions {
-  token: string;
-  path: string;
-  workspaceHandle: string;
-  templateId: string;
-  config: TemplateEditorEmbedConfig;
-}
-
-interface TemplateEditorEmbedConfig {
+export interface TemplateEditorConfig {
   features?: TemplateEditorFeatures | undefined;
   fonts?: FontSource[] | undefined;
   appearance?: AppearanceConfig | undefined;
 }
 
-export type FontSource = CssFontSource | CustomFontSource;
-
-export interface CssFontSource {
-  cssSrc: string;
-}
-
-export interface CustomFontSource {
-  family: string;
-  src: string;
-  weight?:
-    | "100"
-    | "200"
-    | "300"
-    | "400"
-    | "500"
-    | "600"
-    | "700"
-    | "800"
-    | "900";
-  style?: "normal" | "italic" | "oblique";
-  unicodeRange?: string;
-}
-
-export interface AppearanceConfig {
-  fontFamily?: string | undefined;
-  colors?: {
-    primary?: string | undefined;
-    neutral?: string | undefined;
-  };
-}
-
-function createEmbedIframe(options: CreateIframeOptions): HTMLIFrameElement {
-  const {path, token, workspaceHandle, templateId, config} = options;
-
-  const iframe = document.createElement("iframe");
-
-  iframe.style.width = "100%";
-  iframe.style.height = "100%";
-  iframe.style.border = "none";
-
-  const iframeUrl = new URL(path);
-
-  iframeUrl.searchParams.set("token", token);
-  iframeUrl.searchParams.set("config", JSON.stringify(config));
-
-  iframeUrl.hash = `/${workspaceHandle}/templates/${templateId}/edit`;
-
-  iframe.src = iframeUrl.toString();
-
-  return iframe;
-}
-
-function parseTokenPayload(token: string): SdkTokenPayload {
-  const tokenPayloadResult = sdkTokenPayloadSchema.safeParse(
-    JSON.parse(atob(token.split(".")[1] ?? "")),
-  );
-
-  if (!tokenPayloadResult.success) {
-    throw new Error("Invalid token payload");
-  }
-
-  return tokenPayloadResult.data;
-}
-
-function validatePermissions(
-  permissions: SdkPermission[],
-  features: TemplateEditorFeatures,
-): void {
-  const hasPermission = (validPermissionSets: SdkPermission[][]) => {
-    return validPermissionSets.every((permissionSet) =>
-      permissionSet.some((permission) => permissions.includes(permission)),
-    );
-  };
-
-  if (!hasPermission(REQUIRED_PERMISSIONS)) {
-    throw new Error(
-      REQUIRED_PERMISSIONS.map((set) => set.join(" or ")).join(", ") +
-        " permissions are required",
-    );
-  }
-
-  if (
-    features &&
-    typeof features === "object" &&
-    Object.keys(features).length === 0
-  ) {
-    throw new Error(
-      "features cannot be an empty object. Omit the parameter or specify at least one feature.",
-    );
+function validateFeatures(features: TemplateEditorFeatures): void {
+  if (features && Object.keys(features).length === 0) {
+    throw new Error("please specify at least one template editor feature.");
   }
 }
-
-type SdkPermission = z.infer<typeof sdkPermissionSchema>;
-const sdkPermissionSchema = z.enum([
-  "workspace:read",
-  "workspace:write",
-  "template:read",
-  "template:write",
-  "user:read",
-  "user:write",
-]);
-
-type SdkTokenPayload = z.infer<typeof sdkTokenPayloadSchema>;
-const sdkTokenPayloadSchema = z.object({
-  workspaceHandle: z.string(),
-  permissions: z.array(sdkPermissionSchema),
-});
 
 const REQUIRED_PERMISSIONS: SdkPermission[][] = [
   ["workspace:read", "workspace:write"],
