+++ src/calls/embedTemplatePage.ts
@@ -1,5 +1,13 @@
-import * as z from "@zod/mini";
 import type {BadgeSdk} from "../sdk.ts";
+import type {FontSource} from "../helpers/appearance.ts";
+import type {AppearanceConfig} from "../helpers/appearance.ts";
+import {createEmbedIframe} from "../utils/iframe.ts";
+import {parseTokenPayload} from "../helpers/token.ts";
+import {
+  hasPermission,
+  validatePermissions,
+  type SdkPermission,
+} from "../helpers/permissions.ts";
 
 export interface EmbedTemplatePageOptions {
   /**
@@ -17,71 +25,27 @@ export interface EmbedTemplatePageOptions {
 export function embedTemplatePage(
   sdk: BadgeSdk,
   element: HTMLElement,
-  options: EmbedTemplatePageOptions,
+  {templateId, features, fonts, appearance}: EmbedTemplatePageOptions,
 ): void {
-  const {templateId, features, fonts, appearance} = options;
   const {workspaceHandle, permissions} = parseTokenPayload(sdk.token);
 
-  validatePermissions(permissions, features ?? {});
+  validatePermissions(permissions, REQUIRED_PERMISSIONS);
+  validatePermissionsByFeatures(permissions, features ?? {});
 
   const iframe = createEmbedIframe({
     token: sdk.token,
     path: sdk.path,
-    workspaceHandle,
-    templateId,
-    config: {
-      features,
-      fonts,
-      appearance,
-    },
+    pathHash: `/${workspaceHandle}/templates/${templateId}/overview`,
+    config: {features, fonts, appearance},
   });
 
   element.replaceChildren(iframe);
 }
 
-interface CreateIframeOptions {
-  token: string;
-  path: string;
-  workspaceHandle: string;
-  templateId: string;
-  config: TemplateEmbedConfig;
-}
-
-interface TemplateEmbedConfig {
+export interface TemplateEmbedConfig {
   features?: TemplateEmbedFeatures | undefined;
   fonts?: FontSource[] | undefined;
   appearance?: AppearanceConfig | undefined;
-}
-
-export type FontSource = CssFontSource | CustomFontSource;
-
-export interface CssFontSource {
-  cssSrc: string;
-}
-
-export interface CustomFontSource {
-  family: string;
-  src: string;
-  weight?:
-    | "100"
-    | "200"
-    | "300"
-    | "400"
-    | "500"
-    | "600"
-    | "700"
-    | "800"
-    | "900";
-  style?: "normal" | "italic" | "oblique";
-  unicodeRange?: string;
-}
-
-export interface AppearanceConfig {
-  fontFamily?: string | undefined;
-  colors?: {
-    primary?: string | undefined;
-    neutral?: string | undefined;
-  };
 }
 
 export interface TemplateEmbedFeatures {
@@ -91,56 +55,10 @@ export interface TemplateEmbedFeatures {
   campaignEditor?: boolean;
 }
 
-function createEmbedIframe(options: CreateIframeOptions): HTMLIFrameElement {
-  const {path, token, workspaceHandle, templateId, config} = options;
-
-  const iframe = document.createElement("iframe");
-
-  iframe.style.width = "100%";
-  iframe.style.height = "100%";
-  iframe.style.border = "none";
-
-  const iframeUrl = new URL(path);
-
-  iframeUrl.searchParams.set("token", token);
-  iframeUrl.searchParams.set("config", JSON.stringify(config));
-
-  iframeUrl.hash = `/${workspaceHandle}/templates/${templateId}/overview`;
-
-  iframe.src = iframeUrl.toString();
-
-  return iframe;
-}
-
-function parseTokenPayload(token: string): SdkTokenPayload {
-  const tokenPayloadResult = sdkTokenPayloadSchema.safeParse(
-    JSON.parse(atob(token.split(".")[1] ?? "")),
-  );
-
-  if (!tokenPayloadResult.success) {
-    throw new Error("Invalid token payload");
-  }
-
-  return tokenPayloadResult.data;
-}
-
-function validatePermissions(
+function validatePermissionsByFeatures(
   permissions: SdkPermission[],
   features: TemplateEmbedFeatures,
 ): void {
-  const hasPermission = (validPermissionSets: SdkPermission[][]) => {
-    return validPermissionSets.every((permissionSet) =>
-      permissionSet.some((permission) => permissions.includes(permission)),
-    );
-  };
-
-  if (!hasPermission(REQUIRED_PERMISSIONS)) {
-    throw new Error(
-      REQUIRED_PERMISSIONS.map((set) => set.join(" or ")).join(", ") +
-        " permissions are required",
-    );
-  }
-
   const failedPermissions = Object.entries(features)
     .filter(([_, enabled]) => enabled)
     .map(([feature]) => ({
@@ -149,7 +67,10 @@ function validatePermissions(
         REQUIRED_PERMISSIONS_BY_FEATURE[feature as keyof TemplateEmbedFeatures],
     }))
     .filter(({permissions}) => permissions !== undefined)
-    .filter(({permissions}) => !hasPermission(permissions));
+    .filter(
+      ({permissions: featurePermissions}) =>
+        !hasPermission(permissions, featurePermissions),
+    );
 
   if (failedPermissions.length > 0) {
     throw new Error(
@@ -162,26 +83,6 @@ function validatePermissions(
     );
   }
 }
-
-type SdkPermission = z.infer<typeof sdkPermissionSchema>;
-const sdkPermissionSchema = z.enum([
-  "workspace:read",
-  "workspace:write",
-  "template:read",
-  "template:write",
-  "pass:read",
-  "pass:write",
-  "campaign:read",
-  "campaign:write",
-  "user:read",
-  "user:write",
-]);
-
-type SdkTokenPayload = z.infer<typeof sdkTokenPayloadSchema>;
-const sdkTokenPayloadSchema = z.object({
-  workspaceHandle: z.string(),
-  permissions: z.array(sdkPermissionSchema),
-});
 
 const REQUIRED_PERMISSIONS: SdkPermission[][] = [
   ["workspace:read", "workspace:write"],
