+++ src/helpers/appearance.ts
@@ -1,0 +1,30 @@
+export interface AppearanceConfig {
+  fontFamily?: string | undefined;
+  colors?: {
+    primary?: string | undefined;
+    neutral?: string | undefined;
+  };
+}
+
+export type FontSource = CssFontSource | CustomFontSource;
+
+export interface CssFontSource {
+  cssSrc: string;
+}
+
+export interface CustomFontSource {
+  family: string;
+  src: string;
+  weight?:
+    | "100"
+    | "200"
+    | "300"
+    | "400"
+    | "500"
+    | "600"
+    | "700"
+    | "800"
+    | "900";
+  style?: "normal" | "italic" | "oblique";
+  unicodeRange?: string;
+}
